Don't recompile everything on version change
