+++ ts/src/task_force_radio.cpp
@@ -9,6 +9,7 @@
 #include "serverData.hpp"
 #include "Teamspeak.hpp"
 #include "Logger.hpp"
+#include "version.h"
 
 extern struct TS3Functions ts3Functions;
